requirements.txt
streamlit
requests
pillow
huggingface_hub